requirements.txt
pygame-ce
numpy
zengl
git+https://github.com/JiffyRob/SNEK2.git@main